update model structure
